stress_test/requirements.txt
kafka-python-ng==2.2.3
pika==1.3.2
python-dotenv==1.0.1
